pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pkb"
version = "0.10.1"
description = "Private Knowledge Base — multi-LLM conversation organizer"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "psycopg[binary]>=3.1",
    "psycopg_pool>=3.1",
    "chromadb>=0.5",
    "anthropic>=0.40",
    "watchdog>=4.0",
    "alembic>=1.13",
    "sqlalchemy>=2.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
google = ["google-genai>=1.0"]
all-llm = ["openai>=1.0", "google-genai>=1.0"]
web = ["fastapi>=0.110", "uvicorn[standard]>=0.30", "jinja2>=3.1"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.9",
    "openai>=1.0",
    "google-genai>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "httpx>=0.27",
    "pytest-asyncio>=0.24",
]

[project.scripts]
pkb = "pkb.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/pkb"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
